refactor(earn): split button builder components
